Add BTWIN-SERVER deployment configuration
This commit introduces a new playbook section for deploying and configuring the BTWIN-SERVER. It includes tasks for updating the apt cache, installing and configuring Nginx, copying the BTWIN-SERVER directory, and setting up Filebeat configurations for both Linux and Windows. Additionally, it configures Nginx to serve the BTWIN-SERVER and removes the default Nginx site, enhancing the overall deployment process.